refactor(precompiles): B20Factory → SOLID versioned structure (BOP-418)
Mirrors the b20_stablecoin restructure (#3960): splits the monolithic
factory into a shared/append-only dispatch.rs/abi.rs/storage.rs plus a
frozen, self-contained logic/v1.rs, resolved via a new versions.rs that
maps BaseUpgrade -> FactoryVersion -> &dyn Factory. Threads BaseUpgrade
through precompile.rs/provider.rs so future behavior changes ship as a
new version instead of editing shipped bytecode logic in place.
