+++ crates/common/precompiles/src/b20_factory/logic/mod.rs
@@ -1,0 +1,7 @@
+//! Versioned business logic for the B-20 token factory precompile.
+
+mod interface;
+pub use interface::Factory;
+
+mod v1;
+pub use v1::{CommonParams, FactoryV1, TokenCreateParams};
