+++ crates/common/precompiles/src/provider.rs
@@ -204,7 +204,11 @@ impl<S: BasePrecompileSpec> BasePrecompiles<S> {
         // EIP-8130 precompiles alike — uses plain `install`; none is observed, by
         // design, since metrics are scoped to the B-20 token call path.
         if self.spec.upgrade() >= BaseUpgrade::Beryl {
-            B20Factory::install_with_observer(&mut precompiles, observer.clone());
+            B20Factory::install_with_observer(
+                &mut precompiles,
+                self.spec.upgrade(),
+                observer.clone(),
+            );
             BerylLookup::install_with_observer(
                 &mut precompiles,
                 self.spec.upgrade(),
